weed-api/pyproject.toml
[tool.poetry]
name = "weed-api"
version = "0.1.0"
description = ""
authors = ["YuNing Chang <yuning.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.3"
boto3 = "^1.35.47"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.16"
duckdb = "^1.1.3"
polars = "^1.13.1"
pytz = "^2024.2"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.32.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "weed_api.main:start"